requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
attrs==21.2.0
    # via
    #   cattrs
    #   jsii
aws-cdk.aws-events==1.119.0
    # via aws-cdk.aws-s3
aws-cdk.aws-iam==1.119.0
    # via
    #   aws-cdk.aws-events
    #   aws-cdk.aws-kms
    #   aws-cdk.aws-s3
aws-cdk.aws-kms==1.119.0
    # via aws-cdk.aws-s3
aws-cdk.aws-s3==1.119.0
    # via -r requirements.in
aws-cdk.cloud-assembly-schema==1.119.0
    # via
    #   aws-cdk.core
    #   aws-cdk.cx-api
aws_cdk.core==1.119.0
    # via
    #   -r requirements.in
    #   aws-cdk.aws-events
    #   aws-cdk.aws-iam
    #   aws-cdk.aws-kms
    #   aws-cdk.aws-s3
aws-cdk.cx-api==1.119.0
    # via
    #   aws-cdk.aws-kms
    #   aws-cdk.aws-s3
    #   aws-cdk.core
aws-cdk.region-info==1.119.0
    # via
    #   aws-cdk.aws-iam
    #   aws-cdk.core
cattrs==1.7.1
    # via jsii
constructs==3.3.130
    # via
    #   aws-cdk.aws-events
    #   aws-cdk.aws-iam
    #   aws-cdk.aws-kms
    #   aws-cdk.aws-s3
    #   aws-cdk.core
jsii==1.33.0
    # via
    #   aws-cdk.aws-events
    #   aws-cdk.aws-iam
    #   aws-cdk.aws-kms
    #   aws-cdk.aws-s3
    #   aws-cdk.cloud-assembly-schema
    #   aws-cdk.core
    #   aws-cdk.cx-api
    #   aws-cdk.region-info
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk.aws-events
    #   aws-cdk.aws-iam
    #   aws-cdk.aws-kms
    #   aws-cdk.aws-s3
    #   aws-cdk.cloud-assembly-schema
    #   aws-cdk.core
    #   aws-cdk.cx-api
    #   aws-cdk.region-info
    #   constructs
python-dateutil==2.8.2
    # via jsii
six==1.16.0
    # via python-dateutil
typing-extensions==3.10.0.0
    # via jsii
